refactored new event package into goal from UI
