+++ src/main/java/com/karankumar/bookproject/ui/goal/ReadingGoalView.java
@@ -26,7 +26,7 @@ import com.karankumar.bookproject.backend.service.ReadingGoalService;
 import com.karankumar.bookproject.backend.util.StringUtils;
 import com.karankumar.bookproject.backend.util.DateUtils;
 import com.karankumar.bookproject.ui.MainView;
-import com.karankumar.bookproject.ui.events.SaveGoalEvent;
+import com.karankumar.bookproject.ui.goal.events.SaveGoalEvent;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.html.H1;
 import com.vaadin.flow.component.html.H3;
