+++ src/main/java/com/karankumar/bookproject/ui/goal/events/ComponentGoalEvent.java
@@ -15,7 +15,7 @@
     If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.karankumar.bookproject.ui.events;
+package com.karankumar.bookproject.ui.goal.events;
 
 import com.karankumar.bookproject.backend.entity.ReadingGoal;
 import com.karankumar.bookproject.ui.goal.ReadingGoalForm;
